requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
certifi==2024.2.2
click==8.1.7
docutils==0.19
easygui==0.98.3
exceptiongroup==1.2.0
fastapi==0.109.2
fastapi-utils==0.2.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
ifaddr==0.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
libsass==0.23.0
markdown2==2.4.13
MarkupSafe==2.1.5
multidict==6.0.5
nicegui==1.4.19
numpy==1.24.4
opencv-python==4.9.0.80
orjson==3.9.15
pscript==0.7.7
pydantic==1.10.14
pydantic_core==2.16.3
Pygments==2.17.2
pyserial==3.5
python-dotenv==1.0.1
python-engineio==4.9.0
python-multipart==0.0.9
python-socketio==5.11.1
PyYAML==6.0.1
simple-websocket==1.0.0
sniffio==1.3.1
SQLAlchemy==1.4.52
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.29.0
uvloop==0.19.0
vbuild==0.8.2
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
